+++ src/components/YnputCloud/YnputCloud.styled.js
