+++ src/pages/OnBoarding/Step/BoostrapStart.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import * as Styled from '../util/OnBoardingStep.styled'
 
-import YnputConnector from '../../../components/YnputConnect/YnputConnector'
-import * as YnputConnect from '../../../components/YnputConnect/YnputConnect.styled'
+import YnputConnector from '../../../components/YnputCloud/YnputConnector'
+import * as YnputConnect from '../../../components/YnputCloud/YnputCloud.styled'
 
 export const BootstrapStart = ({
   nextStep,
